create tuple producer for symmetry
